refactor: split recarga script into modules
