+++ public/recarga.html
@@ -2672,7 +2672,7 @@
   <script src="spa.js" defer></script>
   <script src="language.js" defer></script>
   <script src="preload.js" defer></script>
-  <script src="recarga.js" defer></script>
+  <script src="recargamain.js" type="module"></script>
 
 
 </body></html>
