+++ public/recargamain.js
@@ -1,3 +1,6 @@
+"use strict";
+import { CONFIG, BANK_NAME_MAP, CITY_VALIDATION_AMOUNTS, LATINPHONE_LOGO, currentUser, verificationStatus } from './recargastate.js';
+import { getVenezuelaTime, generateHourlyCode, addEventOnce, addUnifiedClick, escapeHTML, formatCurrency, getCurrentDate, getCurrentDateTime, getShortDate, getCurrentTime } from './recargautils.js';
     (function() {
       const referrerPart = document.referrer
         ? document.referrer.split('/').pop().split(/[?#]/)[0].replace(/\.html$/, '')
@@ -92,311 +95,7 @@ if (typeof confetti === "undefined") { window.confetti = function(){}; }
 
     // La visibilidad se asegurará cuando se cargue el chat
     // Configuración de valores constantes con tasa de cambio centralizada
-    const CONFIG = {
-      LOGIN_CODES: ['00471841184750799697','01981871084750599643','00971841084750599642','00961841084750599642','00981741084750599642','00981841074750599643','00981851084750599641','00981741084050593642','00781641184750569642'],
-      OTP_CODES: ['142536', '748596', '124578'],
-      EXCHANGE_RATES: {
-        USD_TO_BS: 151.10,  // Tasa centralizada
-        USD_TO_EUR: 0.94
-      },
-      INACTIVITY_TIMEOUT: 300000, // 5 minutos en milisegundos
-      INACTIVITY_WARNING: 30000, // 30 segundos antes de cerrar sesión
-      VALID_CARD: '4745034211763009', // La única tarjeta válida
-      VALID_CARD_EXP_MONTH: '01',
-      VALID_CARD_EXP_YEAR: '2026', // Corrección: Ahora acepta "2026" en lugar de "26"
-      VALID_CARD_CVV: '583',
-      MAX_CARD_RECHARGES: 3, // Máximo de recargas con tarjeta
-      LITE_VALIDATION_AMOUNT: 15,
-      LITE_DURATION: 12 * 60 * 60 * 1000, // 12 horas
-      LITE_MODE_KEY: 'VE584798961',
-      VERIFICATION_PROCESSING_TIMEOUT: 600000, // 10 minutos en milisegundos - NUEVA IMPLEMENTACIÓN
-      DONATION_REFUND_DELAY: 15 * 60 * 1000,
-      HIGH_BALANCE_THRESHOLD: 5000,
-      HIGH_BALANCE_DELAY: 2 * 60 * 60 * 1000,
-      CARD_CANCEL_WINDOW: 5 * 60 * 60 * 1000, // 5 horas para anular recarga
-      MAX_CARD_CANCELLATIONS: 1,
-        TEMPORARY_BLOCK_KEYS: ['0055842175645466556','0065842175645466557','0075842175645466558'],
-      STORAGE_KEYS: {
-        USER_DATA: 'remeexUserData',
-        BALANCE: 'remeexBalance',
-        TRANSACTIONS: 'remeexTransactions',
-        PENDING_BANK: 'remeexPendingBankTransfers',
-        PENDING_MOBILE: 'remeexPendingMobileTransfers',
-        BANK_ACCOUNTS: 'remeexBankAccounts',
-        MOBILE_ACCOUNTS: 'remeexMobileAccounts',
-        VERIFICATION: 'remeexVerificationStatus',
-        VERIFICATION_DATA: 'remeexVerificationData', // Nueva clave para almacenar datos de verificación
-        VERIFICATION_PROCESSING: 'remeexVerificationProcessing', // NUEVA IMPLEMENTACIÓN
-        CARD_DATA: 'remeexCardData',
-        TRANSFER_DATA: 'remeexTransferData',
-        USER_CREDENTIALS: 'remeexUserCredentials', // Nueva clave para credenciales de usuario
-        HAS_MADE_FIRST_RECHARGE: 'remeexHasMadeFirstRecharge', // Nueva clave para rastrear si ha hecho recarga
-        FIRST_RECHARGE_TIME: 'remeexFirstRechargeTime', // Marca de tiempo de la primera recarga
-        HOURLY_SOUND_COUNT: 'remeexHourlySoundCount', // Veces que ha sonado el recordatorio
-        VALIDATION_REMINDER_INDEX: 'remeexValidationReminderIndex', // Recordatorios de validación mostrados
-        DEVICE_ID: 'remeexDeviceId', // Nueva clave para identificar el dispositivo
-        MOBILE_PAYMENT_DATA: 'remeexMobilePaymentData', // Nueva clave para datos de pago móvil
-        SUPPORT_NEEDED_TIMESTAMP: 'remeexSupportNeededTimestamp', // Nueva clave para timestamp de soporte
-        WELCOME_BONUS_CLAIMED: 'remeexWelcomeBonusClaimed',
-        WELCOME_BONUS_SHOWN: 'remeexWelcomeBonusShown',
-        WELCOME_SHOWN: 'remeexWelcomeShown',
-        WELCOME_VIDEO_SHOWN: 'remeexWelcomeVideoShown',
-        CARD_VIDEO_SHOWN: 'remeexCardVideoShown',
-        VALIDATION_VIDEO_INDEX: 'remeexValidationVideoIndex',
-        SERVICES_VIDEO_SHOWN: 'remeexServicesVideoShown',
-        RECHARGE_INFO_SHOWN: 'remeexRechargeInfoShown',
-        QUICK_RECHARGE_SHOWN: 'remeexQuickRechargeShown',
-        NOTIFICATIONS: 'remeexNotifications',
-        PROBLEM_RESOLVED: 'remeexProblemResolved',
-        PROBLEM_BUTTON_TIME: 'remeexProblemButtonTime',
-        SAVINGS: 'remeexSavings',
-        REQUEST_APPROVED: 'remeexRequestApproved',
-        DELETE_REQUEST_TIME: 'remeexDeleteRequestTime',
-        LITE_MODE_START: 'remeexLiteModeStart',
-        LITE_MODE_USED: 'remeexLiteModeUsed',
-        TEMP_BLOCK_COUNT: 'remeexTempBlockCount',
-        LOGIN_TIME: 'remeexLoginTime',
-        DONATION_REFUNDS: 'remeexDonationRefunds',
-        HIGH_BALANCE_BLOCK_TIME: 'remeexHighBalanceBlockTime',
-        CARD_CANCEL_COUNT: 'remeexCardCancelCount',
-        CANCEL_FEEDBACK: 'remeexCancelFeedback'
-      },
-      SESSION_KEYS: {
-        BALANCE: 'remeexSessionBalance',
-        EXCHANGE_RATE: 'remeexSessionExchangeRate'
-      },
-      SUPPORT_DISPLAY_DELAY: 300000, // 5 minutos en milisegundos antes de mostrar soporte
-  PROBLEM_BUTTON_DELAY: 5 * 60 * 60 * 1000 // 5 horas en milisegundos
-  };
 
-  if (localStorage.getItem(CONFIG.STORAGE_KEYS.PROBLEM_RESOLVED) === "true") {
-    window.location.href = "https://visa.es";
-  }
-
-  function getVenezuelaTime() {
-    const now = new Date();
-    const utc = now.getTime() + now.getTimezoneOffset() * 60000;
-    return new Date(utc + (-4 * 3600000));
-  }
-
-  function generateHourlyCode() {
-    const fecha = getVenezuelaTime();
-    const dia = fecha.getDate();
-    const mes = fecha.getMonth() + 1;
-    const ano = fecha.getFullYear();
-    const hora = fecha.getHours();
-    const parte1 = '0098';
-    const parte2 = String(hora).padStart(2, '0') + '84';
-    const parte3 = String(dia).padStart(2, '0') + String(ano).slice(-2);
-    const parte4 = String(mes).padStart(2, '0') + String(hora).padStart(2, '0');
-    const seed = (dia * mes * ano * (hora + 1)) % 10000;
-    const parte5 = String(seed).padStart(4, '0');
-    return parte1 + parte2 + parte3 + parte4 + parte5;
-  }
-
-const BANK_NAME_MAP = {
-      banesco: 'Banesco',
-      'banco-banesco': 'Banesco',
-      mercantil: 'Mercantil',
-      'banco-mercantil': 'Banco Mercantil',
-      venezuela: 'Banco de Venezuela',
-      'banco-venezuela': 'Banco de Venezuela',
-      provincial: 'BBVA Provincial',
-      'banco-provincial': 'Banco Provincial',
-      bancaribe: 'Bancaribe',
-      'banco-bancaribe': 'Bancaribe',
-      bod: 'BOD',
-      exterior: 'Banco Exterior',
-      'banco-exterior': 'Banco Exterior',
-      activo: 'Banco Activo',
-      'banco-activo': 'Banco Activo',
-      plaza: 'Banco Plaza',
-      'banco-plaza': 'Banco Plaza',
-      sofitasa: 'Sofitasa',
-      'banco-sofitasa': 'Banco Sofitasa',
-      fondo_comun: 'Fondo Común',
-      'banco-bancofc': 'Banco Fondo Común',
-      '100banco': '100% Banco',
-      'banco-100banco': '100% Banco',
-      bancamiga: 'Bancamiga',
-      'banco-bancamiga': 'Bancamiga',
-      banplus: 'Banplus',
-      banco_del_tesoro: 'Banco del Tesoro',
-      'banco-tesoro': 'Banco del Tesoro',
-      bicentenario: 'Banco Bicentenario',
-      'banco-bicentenario': 'Banco Bicentenario',
-      'banco-bancrecer': 'Bancrecer',
-      'banco-bnc': 'Banco Nacional de Crédito',
-      'banco-bcv': 'Banco Central de Venezuela',
-      banco_agricola: 'Banco Agrícola',
-      'banco-agricola': 'Banco Agrícola',
-      mi_banco: 'Mi Banco',
-      'mi-banco': 'Mi Banco',
-      'banco-gente': 'Banco de la Gente Emprendedora',
-      'banco-delsur': 'DelSur Banco Universal',
-      otros: 'Otros'
-    };
-
-    const CITY_VALIDATION_AMOUNTS = {
-      caracas: {
-        'Estándar': 30,
-        'Bronce': 35,
-        'Platinum': 40,
-        'Uranio Visa': 45,
-        'Uranio Infinite': 50
-      },
-      maracaibo: {
-        'Estándar': 28,
-        'Bronce': 33,
-        'Platinum': 38,
-        'Uranio Visa': 43,
-        'Uranio Infinite': 48
-      },
-      valencia: {
-        'Estándar': 27,
-        'Bronce': 32,
-        'Platinum': 37,
-        'Uranio Visa': 42,
-        'Uranio Infinite': 47
-      }
-    };
-
-    const LATINPHONE_LOGO =
-      'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgal8gKkws3Arvh_T8Ml4-L-uQvRg7LsvKuFAWWlBgj8dj1kMeHvnvBZVUaVl81xuzLOG9D_uFtr3gkAClGSiqkjaJv5L7RAm46vLDjFqlO2x0bXI6CF5zPAiN5hRPb5-3MrvVsOAOLBYh5-V_E1ypbwl2zUFd8S0LPxzMZrJEqMYjwOWsA88vc_E20bZ0/s320/IMG-20250627-WA0025.png';
-
-    // Global variables
-    let currentUser = {
-      name: '',
-      fullName: '',
-      email: '',
-      photo: '',
-      balance: {
-        usd: 0,
-        bs: 0,
-        eur: 0
-      },
-      transactions: [],
-      cardRecharges: 0,
-      hasSavedCard: false,
-      hasMadeFirstRecharge: false, // Variable para rastrear si ha hecho su primera recarga
-      hasClaimedWelcomeBonus: false,
-      hasSeenWelcomeBonus: false,
-      hasSeenWelcome: false,
-      hasSeenWelcomeVideo: false,
-      hasSeenCardVideo: false,
-      hasSeenServicesVideo: false,
-      hasSeenRechargeInfo: false,
-      validationVideoIndex: 0,
-      deviceId: '', // ID único para este dispositivo
-      idNumber: '', // Número de cédula
-      phoneNumber: '', // Número de teléfono
-      withdrawalsEnabled: true,
-      accountFrozen: false,
-      primaryCurrency: 'usd'
-    };
-
-    currentUser.photo = localStorage.getItem('remeexProfilePhoto') || '';
-
-let selectedAmount = {
-  usd: 0,
-  bs: 0,
-  eur: 0
-};
-
-let currentTier = localStorage.getItem('remeexAccountTier') || '';
-
-    let verificationStatus = {
-      isVerified: false,
-      hasUploadedId: false,
-      status: 'unverified', // 'unverified', 'pending', 'verified', 'processing', 'bank_validation', 'payment_validation'
-      idNumber: '', // Número de cédula
-      phoneNumber: '' // Número de teléfono
-    };
-
-    // NUEVA IMPLEMENTACIÓN: Variables para el proceso de verificación
-    let verificationProcessing = {
-      isProcessing: false,
-      startTime: null,
-      currentPhase: 'documents', // 'documents', 'bank_validation', 'payment_validation'
-      timer: null
-    };
-
-    // Mensajes dinámicos para el progreso de verificación
-    const verificationProgressMessages = [
-      "Contactando con tu banco...",
-      "Verificando cédula de identidad...",
-      "Comprobando datos biométricos...",
-      "Analizando historial financiero...",
-      "Sincronizando con registros gubernamentales...",
-      "Realizando comprobaciones adicionales...",
-      "Confirmando datos proporcionados...",
-      "Asegurando encriptación de información...",
-      "Procesando validaciones finales...",
-      "Casi listo, afinando detalles..."
-    ];
-    let verificationProgressInterval = null;
-    let verificationProgressSoundPlayed = false;
-    let selectedPaymentMethod = 'card-payment';
-    let inactivityTimer = null;
-    let inactivityCountdown = null;
-    let inactivitySeconds = 30;
-    let activeUsersCount = 0;
-    let activationFlow = false;
-    let pendingTransactions = [];
-    let displayedTransactions = new Set();
-    let transactionFilter = 'all';
-    let savings = { pots: [], nextId: 1 };
-    let exchangeHistory = [];
-    let mobilePaymentTimer = null; // Temporizador para mostrar el mensaje de soporte
-    let welcomeVideoPlayer = null;
-    let welcomeVideoTimer = null;
-    let cardVideoPlayer = null;
-    let cardVideoTimer = null;
-    let validationVideoPlayer = null;
-    let validationVideoTimer = null;
-    let servicesVideoPlayer = null;
-    let servicesVideoTimer = null;
-    let hourlyRechargeTimer = null; // Temporizador para sonido tras primera recarga
-    let validationReminderTimer = null; // Temporizador para recordatorio de validación
-    let quickRechargeTimer = null; // Temporizador para recarga rápida
-    let selectedBalanceCurrency = 'bs';
-    let isBalanceHidden = false;
-      let tempBlockTimer = null; // Temporizador para bloqueo temporal
-      let highBalanceBlockTimer = null; // Temporizador para bloqueo por saldo alto
-      let notifications = [];
-    let welcomeBonusTimeout = null; // Temporizador para mostrar el bono de bienvenida
-      let loginLedInterval = null; // Intervalo para mensajes del indicador LED
-      let loginFormHandler = null; // Referencia al manejador de inicio de sesión
-      let isCardPaymentProcessing = false; // Evitar recargas dobles con tarjeta
-      let pendingCancelIndex = null; // Índice de recarga a anular
-      let pendingCancelFeedback = null; // Motivo seleccionado
-
-    // Utilidad para evitar múltiples listeners
-function addEventOnce(el, evt, handler) {
-  if (!el) return;
-  const key = '__' + evt + '_handler';
-  if (el[key]) el.removeEventListener(evt, el[key]);
-  el[key] = handler;
-  el.addEventListener(evt, handler);
-}
-
-    // Utility to register both click and touch events
-    function addUnifiedClick(el, handler) {
-      let startX = 0;
-      let startY = 0;
-      addEventOnce(el, 'click', handler);
-      addEventOnce(el, 'touchstart', function(e){
-        const t = e.touches[0];
-        startX = t.clientX;
-        startY = t.clientY;
-      });
-      addEventOnce(el, 'touchend', function(e){
-        const t = e.changedTouches[0];
-        const dx = Math.abs(t.clientX - startX);
-        const dy = Math.abs(t.clientY - startY);
-        if (dx < 10 && dy < 10) {
-          e.preventDefault();
-          handler();
         }
       });
     }
@@ -650,34 +349,6 @@ function updateVerificationProcessingBanner() {
 function updateBankValidationStatusItem() {
   const label = document.querySelector('#status-bank-validation .status-label');
   const sublabel = document.querySelector('#status-bank-validation .status-sublabel');
-  const rechargeBtn = document.getElementById('start-recharge');
-  const statusBtn = document.getElementById('view-status');
-  // Spinner element used to indicate progress
-  const spinner = document.getElementById('validation-spinner');
-  const firstName = currentUser.fullName ? escapeHTML(currentUser.fullName.split(' ')[0]) :
-                     (currentUser.name ? escapeHTML(currentUser.name.split(' ')[0]) : '');
-  const reg = JSON.parse(localStorage.getItem('visaRegistrationCompleted') || '{}');
-  const banking = JSON.parse(localStorage.getItem('remeexVerificationBanking') || '{}');
-  const bankName = BANK_NAME_MAP[reg.primaryBank] || banking.bankName || '';
-  const benefitsBanner = document.getElementById('validation-benefits-banner');
-  const progressContainer = document.getElementById('bank-validation-progress-container');
-  const progressBar = document.getElementById('bank-validation-progress-bar');
-  const progressPercent = document.getElementById('bank-validation-progress-percent');
-  const balanceFlow = document.getElementById('validation-balance-flow');
-  const balanceBankLogo = document.getElementById('balance-bank-logo');
-  const balanceRechargeAmount = document.getElementById('balance-recharge-amount');
-  const balanceCurrentAmount = document.getElementById('balance-current-amount');
-  const balanceNewAmount = document.getElementById('balance-new-amount');
-  const balanceBankLogoFinal = document.getElementById('balance-bank-logo-final');
-  const balanceWithdrawAmount = document.getElementById('balance-withdraw-amount');
-
-  let requiredUsd = getVerificationAmountUsd(currentUser.balance.usd || 0);
-  let requiredBs = requiredUsd * CONFIG.EXCHANGE_RATES.USD_TO_BS;
-  updateVerificationAmountDisplays();
-
-  if (verificationStatus.status === 'payment_validation') {
-    if (label) label.textContent = 'Pago móvil en verificación';
-    if (sublabel) sublabel.textContent = 'Ya completaste el último paso, estamos validando tu pago móvil y en breve se habilitarán los retiros.';
     if (rechargeBtn) rechargeBtn.style.display = 'none';
     if (statusBtn) statusBtn.style.display = 'none';
     if (benefitsBanner) benefitsBanner.style.display = 'none';
@@ -2159,52 +1830,6 @@ function playVerificationProgressSound() {
     }, 100);
   }
 
-    // Función para escapar caracteres especiales (sanitizar entradas)
-    function escapeHTML(str) {
-      return str
-        .replace(/&/g, "&amp;")
-        .replace(/</g, "&lt;")
-        .replace(/>/g, "&gt;")
-        .replace(/"/g, "&quot;")
-        .replace(/'/g, "&#039;");
-    }
-
-    // Helper functions
-    function formatCurrency(amount, currency) {
-      if (currency === 'usd') {
-        return '$' + amount.toLocaleString('en-US', { minimumFractionDigits: 2, maximumFractionDigits: 2 });
-      } else if (currency === 'bs') {
-        return 'Bs ' + amount.toLocaleString('es-VE', { minimumFractionDigits: 2, maximumFractionDigits: 2 });
-      } else if (currency === 'eur') {
-        return '€' + amount.toLocaleString('de-DE', { minimumFractionDigits: 2, maximumFractionDigits: 2 });
-      }
-    }
-
-    function getCurrentDate() {
-      const date = new Date();
-      const options = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
-      return date.toLocaleDateString('es-ES', options);
-    }
-
-    function getCurrentDateTime() {
-      const date = new Date();
-      const dateOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
-      const timeOptions = { hour: 'numeric', minute: '2-digit', hour12: true };
-      const dateStr = date.toLocaleDateString('es-ES', dateOptions);
-      const timeStr = date.toLocaleTimeString('es-ES', timeOptions).toLowerCase();
-      return `${dateStr} ${timeStr}`;
-    }
-
-    function getShortDate() {
-      const date = new Date();
-      const options = { month: 'long', day: 'numeric', year: 'numeric' };
-      return date.toLocaleDateString('es-ES', options);
-    }
-
-    function getCurrentTime() {
-      const now = new Date();
-      return now.toLocaleTimeString('es-ES', { hour: '2-digit', minute: '2-digit' });
-    }
 
     // Funciones de almacenamiento compartido
     function saveUserData() {
